src/image2caption/requirements.txt
google-generativeai
datasets
aiohttp
pillow
pandas
tqdm
python-dotenv
huggingface-hub
